@@ -1,4 +1,4 @@
-class CreateLanguages < ActiveRecord::Migration
+class CreateLanguages < ActiveRecord::Migration[4.2]
   def change
     create_table :languages do |t|
       t.string :iso
