@@ -1,4 +1,4 @@
-class AddTokenToApp < ActiveRecord::Migration
+class AddTokenToApp < ActiveRecord::Migration[4.2]
   def change
     add_column :apps, :token, :string
   end
